Added proxy setting
Added a proxy setting to package.json so that the frontend can pass
requests for the REST API through to the backend port.  Changed the
URLs used in the fetch() calls in BookmarkAPI to be relative when run
from a browser, but absolute when running under npm test (since
isomorphic-fetch requires URLs to be absolute).